% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterClusters.R
\name{filterClusters}
\alias{filterClusters}
\title{Mapped cluster filter}
\usage{
filterClusters(table, min_main_cluster_freq = 0.6)
}
\arguments{
\item{table}{data.table with original cluster information}

\item{min_main_cluster_freq}{Numeric between 0 and 1. Minimum frequency of a
transition group id matched to the same cluster. Default is 0.6.}
}
\value{
data.table Same as input table but including extra columns about
run coverage information and filtered by \code{min_main_cluster_freq}.
}
\description{
Filter mapped clusters by following criteria:
\itemize{
 \item Cluster must be main cluster of the transition group.
 \item Main cluster frequency must be >= min_main_cluster_freq.
 \item Main cluster should not be zero (non-defined).
}
}
\examples{
data <- fullDataMapped
filteredData <- filterData(data,
                           min_main_cluster_freq = 0.6)
}
