% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterData.R
\name{filterData}
\alias{filterData}
\title{Coverage filter}
\usage{
filterData(table, min_run_coverage = 0.6)
}
\arguments{
\item{table}{data.table with original cluster information}

\item{min_run_coverage}{Numeric between 0 and 1. Minimum coverage of a
transition group id across all runs. Default is 0.6.}
}
\value{
data.table Same as input table but including extra columns about
run coverage information and filtered by \code{min_run_coverage}.
}
\description{
Filter input data by removing transition groups with a detection frequency < min_run_coverage.
}
\examples{
data <- exampleData
filteredData <- filterData(data,
                           min_run_coverage = 0.6)
}
