% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMatrix.R
\name{generateMatrix}
\alias{generateMatrix}
\title{Generate final output matrix}
\usage{
generateMatrix(rawInput, processedData)
}
\arguments{
\item{rawInput}{data.table with original cluster information.}

\item{processedData}{data.table with mapped and filtered clutsre information.}
}
\value{
data.table Same as input table but including extra columns with
new_cluster information.
}
\description{
Merge input data with mapped and filtered results.
}
\examples{
data <- exampleData
filteredData <- filterData(data,
                           min_run_coverage = 0.6)
mappedData <- mapClusters(filteredData)
mappedDataFiltered <- filterClusters(mappedData,
                                     min_main_cluster_freq = 0.6)
finalData <- generateMatrix(rawInput = data,
                            processedData = mappedDataFiltered)

}
