% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapClusters.R
\name{mapClusters}
\alias{mapClusters}
\title{Cluster mapping function}
\usage{
mapClusters(table, sort_RT = FALSE)
}
\arguments{
\item{table}{data.table with original cluster information}

\item{sort_RT}{Logical whether to sort the the original Clusters of test runs by retention time.
Default is \code{FALSE}.}
}
\value{
data.table Same as input table but including extra columns about
mapping information. The "new_cluster" column contains the mapped cluster information.
}
\description{
Map clusters by common transition groups.
}
\examples{
data <- exampleDataFiltered
dataMapped <- mapClusters(data)

}
